requirements.txt
langchain>=0.0.267
langchain-community>=0.0.5
langchain-ollama==0.0.3
ollama>=0.1.2
chromadb>=0.4.13
duckduckgo-search>=3.9.9
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
sentence-transformers>=2.2.2
openai>=1.3.0
tiktoken>=0.5.1 